+++ packages/atom-react-redux/src/common/actions.ts
@@ -1,7 +1,11 @@
 import { createAction } from "@reduxjs/toolkit";
-import type { RaiseErrorInfo } from "@hwndmaster/atom-web-core";
+import type { LoadingParam, LoadingTarget, LoadingTargetPayload, RaiseErrorInfo } from "@hwndmaster/atom-web-core";
 
-export const showLoader = createAction<number>("common/showLoader");
-export const hideLoader = createAction<number>("common/hideLoader");
+function prepareLoadingTarget(target: LoadingTarget, param?: LoadingParam): { payload: LoadingTargetPayload } {
+    return { payload: param == null ? target : { target, param } };
+}
+
+export const showLoader = createAction("common/showLoader", prepareLoadingTarget);
+export const hideLoader = createAction("common/hideLoader", prepareLoadingTarget);
 export const hideAllLoaders = createAction<void>("common/hideAllLoaders");
 export const raiseError = createAction<RaiseErrorInfo>("common/raiseError");