+++ packages/atom-react-redux/src/common/index.ts
@@ -2,3 +2,4 @@ export * as Actions from "./actions";
 export { default as Reducers } from "./reducers";
 export * as Sagas from "./sagas";
 export type { default as CommonState } from "./state";
+export { buildLoadingKey, loadingKeyFromPayload } from "./loadingKey";