+++ packages/atom-web-core/src/index.ts
@@ -19,7 +19,7 @@ export {
 } from "./helpers";
 
 export { BaseLoadingTargets } from "./loadingTargets";
-export type { LoadingTarget } from "./loadingTargets";
+export type { LoadingTarget, LoadingParam, ParametrizedLoadingTarget, LoadingTargetPayload } from "./loadingTargets";
 
 export { setNotificationService, getNotificationService } from "./notifications";
 export type { NotificationService } from "./notifications";