+++ packages/atom-api-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hwndmaster/atom-api-core
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @hwndmaster/atom-web-core@0.1.6
+
 ## 0.1.5
 
 ### Patch Changes